+++ src/main/java/net/spy/memcached/ops/MutatorOperation.java
@@ -23,5 +23,5 @@ public interface MutatorOperation extends KeyedOperation {
   /**
    * Get the expiration to set in case of a new entry.
    */
-  int getExpiration();
+  long getExpiration();
 }
