+++ src/test/manual/net/spy/memcached/collection/attribute/UnReadableSetTest.java
@@ -71,7 +71,7 @@ class UnReadableSetTest extends BaseIntegrationTest {
 
       assertEquals(Long.valueOf(0), attr.getCount());
       assertEquals(Long.valueOf(4000), attr.getMaxCount());
-      assertEquals(Integer.valueOf(0), attr.getExpireTime());
+      assertEquals(Long.valueOf(0), attr.getExpireTime());
       assertFalse(attr.getReadable());
 
       // insert an item
@@ -118,7 +118,7 @@ class UnReadableSetTest extends BaseIntegrationTest {
 
       assertEquals(Long.valueOf(0), attr.getCount());
       assertEquals(Long.valueOf(4000), attr.getMaxCount());
-      assertEquals(Integer.valueOf(0), attr.getExpireTime());
+      assertEquals(Long.valueOf(0), attr.getExpireTime());
       assertTrue(attr.getReadable());
 
       // insert an item
